+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 1.0.1
 
-- "Fixed" `EnableDebug` somehow being true by default, despite it CLEARLY being declared as false?? I have no clue what happened here...
+- "Fixed" config entry `EnableDebug` somehow being true by default, despite it CLEARLY being declared as false?? I have no clue what happened here...
 - Updated Discord URLs in README to link to this mod's dedicated thread now.
 
 ### 1.0.0 - Initial release 🧟‍♀️